+++ src/model/es/model/esConnectionNode.ts
@@ -63,7 +63,7 @@ export class EsConnectionNode extends Node {
 
 
     newQuery() {
-        QueryUnit.showSQLTextDocument(this,EsTemplate.query,`${this.host}.es`)
+        QueryUnit.showSQLTextDocument(this,EsTemplate.query,`${this.host.replace(/^(http|https):/,'').replace(/\//g,'')}.es`)
     }
 
     async getChildren(): Promise<Node[]> {
@@ -73,7 +73,7 @@ export class EsConnectionNode extends Node {
     }
 
     public copyName() {
-        Util.copyToBoard(this.host)
+        Util.copyToBoard(this.usingSSH ? this.ssh.host : this.host)
     }
 
     public async deleteConnection(context: ExtensionContext) {