+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # CHANGELOG
 
-# 4.5.2 2021-1-21
+# 4.5.7 2022-2-6
+
+- Support connect to redis with username.
+- Auto scroll redis terminal.
+- Copy ssh host if using ssh.
+- Fix open es query not active connection #450.
+- Better query file for table.
+
+# 4.5.2 2022-1-21
 
 - Add more error tolerance when connecting.
 - Support show redis server version.
@@ -14,7 +22,7 @@
 - Fix sqlite complection fail.
 - Other bugs fix.
 
-# 4.5.0 2021-1-7
+# 4.5.0 2022-1-7
 
 - Support connect to clickhouse.
 - Show results in vertical direction.